+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 * `Linna\Filter\Rules\Ip` check for valid ipv4 and ipv6 address
-* `Linna\Filter\Rules\IPv4Range` check if provided ipv4 is in CIDR range
+* `Linna\Filter\Rules\IPRange` check if provided ipv4/ipv6 is in CIDR range
 * Namespace for tests
 
 ### Changed